refactor: move `internal_types` into util
