+++ src/components/NavStepper.tsx
@@ -3,8 +3,8 @@
 // to the `slots.stepIcon` prop, and defining a simple function inline is the most
 // straightforward way to achieve this. The performance impact is minimal in this case.
 import { Stepper, Step, StepLabel, StepConnector, useTheme, SvgIcon } from '@mui/material';
-import { View } from '../../internal_types';
 import { steps } from '../utils/constants';
+import { View } from '../utils/internal_types';
 
 function NavStepper({ currentView }: { currentView: View }) {
   const theme = useTheme();
