+++ src/hooks/useColumnsMetadata.test.ts
@@ -1,7 +1,7 @@
 import { renderHook } from '@testing-library/react';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
-import { DataType, type Columns } from '../../internal_types';
 import { useColumns, useStandardizedVariables } from '../stores/data';
+import { DataType, type Columns } from '../utils/internal_types';
 import { useColumnsMetadata } from './useColumnsMetadata';
 
 vi.mock('../stores/data', () => ({
