+++ src/hooks/useMultiColumnMeasureTabs.test.ts
@@ -1,6 +1,6 @@
 import { act, renderHook } from '@testing-library/react';
 import { describe, it, expect, beforeEach, vi } from 'vitest';
-import type { StandardizedVariable } from '../../internal_types';
+import type { StandardizedVariable } from '../utils/internal_types';
 import { useMultiColumnMeasureTabs } from './useMultiColumnMeasureTabs';
 import { useMultiColumnMeasureVariables } from './useMultiColumnMeasureVariables';
