+++ src/hooks/useMultiColumnMeasureTabs.ts
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react';
-import { StandardizedVariable } from '../../internal_types';
+import { StandardizedVariable } from '../utils/internal_types';
 import { useMultiColumnMeasureVariables } from './useMultiColumnMeasureVariables';
 
 export interface ActiveMultiColumnVariable {
