+++ src/hooks/useTermOptions.test.ts
@@ -1,7 +1,7 @@
 import { renderHook } from '@testing-library/react';
 import { describe, it, expect, beforeEach, vi } from 'vitest';
 import { useStandardizedTerms } from '~/stores/data';
-import type { StandardizedTerms } from '../../internal_types';
+import type { StandardizedTerms } from '../utils/internal_types';
 import { useTermOptions } from './useTermOptions';
 
 vi.mock('~/stores/data', () => ({
