+++ src/hooks/useValueAnnotationColumns.test.ts
@@ -1,8 +1,8 @@
 import { renderHook } from '@testing-library/react';
 import { describe, it, expect, beforeEach, vi } from 'vitest';
 import { useColumns, useStandardizedVariables } from '~/stores/data';
-import type { Columns, StandardizedVariables } from '../../internal_types';
-import { DataType, VariableType } from '../../internal_types';
+import type { Columns, StandardizedVariables } from '../utils/internal_types';
+import { DataType, VariableType } from '../utils/internal_types';
 import { useValueAnnotationColumns } from './useValueAnnotationColumns';
 
 vi.mock('~/stores/data', () => ({
